# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Verdure Plant Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Category {
  id: number;
  name: string;
  nameFr: string;
  nameAr: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  nameFr: string;
  nameAr: string;
  slug: string;
  description: string;
  descriptionFr: string;
  descriptionAr: string;
  price: number;
  stock: number;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  category?: Category;
  featured: boolean;
  careLevel?: string;
  lightRequirement?: string;
  wateringFrequency?: string;
  /** @nullable */
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateProductInput {
  name: string;
  nameFr: string;
  nameAr: string;
  description: string;
  descriptionFr: string;
  descriptionAr: string;
  price: number;
  stock: number;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  featured?: boolean;
  careLevel?: string;
  lightRequirement?: string;
  wateringFrequency?: string;
}

export interface UpdateProductInput {
  name?: string;
  nameFr?: string;
  nameAr?: string;
  description?: string;
  descriptionFr?: string;
  descriptionAr?: string;
  price?: number;
  stock?: number;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  featured?: boolean;
  careLevel?: string;
  lightRequirement?: string;
  wateringFrequency?: string;
}

export interface CreateCategoryInput {
  name: string;
  nameFr: string;
  nameAr: string;
  description?: string;
  image?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UpdateProfileInput {
  name?: string;
  phone?: string;
  address?: string;
  city?: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  status: string;
  total: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  paymentMethod: string;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface CreateOrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  paymentMethod: string;
  items: OrderItemInput[];
}

export interface UpdateOrderStatusInput {
  status: string;
}

export interface OrdersResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export interface WishlistItem {
  id: number;
  userId: number;
  productId: number;
  product: Product;
  createdAt: string;
}

export interface WishlistInput {
  productId: number;
}

export interface Review {
  id: number;
  userId: number;
  productId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  userName: string;
  createdAt: string;
}

export interface CreateReviewInput {
  rating: number;
  comment?: string;
}

export interface NewsletterInput {
  email: string;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
  orders: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalCustomers: number;
  pendingOrders: number;
  lowStockProducts: number;
  recentOrders: Order[];
  monthlyRevenue: MonthlyRevenue[];
  ordersByStatus: StatusCount[];
}

export interface CustomerSummary {
  id: number;
  name: string;
  email: string;
  orderCount: number;
  totalSpent: number;
  createdAt: string;
}

export type GetProductsParams = {
  page?: number;
  limit?: number;
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: string;
};

export type GetOrdersParams = {
  status?: string;
  page?: number;
};